feat: rewrite in nuxtjs
