+++ website/.npmrc
@@ -1,2 +1,2 @@
-engine-strict=true
-resolution-mode=highest
+shamefully-hoist=true
+strict-peer-dependencies=false
