+++ website/README.md
@@ -1,38 +1,63 @@
-# create-svelte
+# Nuxt 3 Minimal Starter
 
-Everything you need to build a Svelte project, powered by [`create-svelte`](https://github.com/sveltejs/kit/tree/master/packages/create-svelte).
+Look at the [Nuxt 3 documentation](https://nuxt.com/docs/getting-started/introduction) to learn more.
 
-## Creating a project
+## Setup
 
-If you're seeing this, you've probably already done this step. Congrats!
+Make sure to install the dependencies:
 
 ```bash
-# create a new project in the current directory
-npm create svelte@latest
+# npm
+npm install
 
-# create a new project in my-app
-npm create svelte@latest my-app
+# pnpm
+pnpm install
+
+# yarn
+yarn install
 ```
 
-## Developing
+## Development Server
 
-Once you've created a project and installed dependencies with `npm install` (or `pnpm install` or `yarn`), start a development server:
+Start the development server on `http://localhost:3000`:
 
 ```bash
+# npm
 npm run dev
 
-# or start the server and open the app in a new browser tab
-npm run dev -- --open
+# pnpm
+pnpm run dev
+
+# yarn
+yarn dev
 ```
 
-## Building
+## Production
 
-To create a production version of your app:
+Build the application for production:
 
 ```bash
+# npm
 npm run build
+
+# pnpm
+pnpm run build
+
+# yarn
+yarn build
 ```
 
-You can preview the production build with `npm run preview`.
+Locally preview production build:
 
-> To deploy your app, you may need to install an [adapter](https://kit.svelte.dev/docs/adapters) for your target environment.
+```bash
+# npm
+npm run preview
+
+# pnpm
+pnpm run preview
+
+# yarn
+yarn preview
+```
+
+Check out the [deployment documentation](https://nuxt.com/docs/getting-started/deployment) for more information.
