+++ website/pages/index.vue
@@ -1,0 +1,25 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <main>
+    <h1>Minecraft Crowd Control</h1>
+
+    <p>
+      Minecraft <a href="https://crowdcontrol.live/">Crowd Control</a> is a Minecraft mod that allows your livestream
+      viewers on any platform to interact with your game using tips, bits, channel points, charitable donations, and more.
+      It can run as a client mod or as a server plugin.
+    </p>
+
+    <h2>Installation</h2>
+    <p>
+      For help installing the mod, see the <NuxtLink to="/setup">dynamic setup guide</NuxtLink>.
+    </p>
+
+    <h2>Credits</h2>
+    <p>
+      This mod was created by <a href="https://qixils.dev/">Lexi</a>
+      with the support of <a href="https://crowdcontrol.live">Crowd Control</a>.
+    </p>
+  </main>
+</template>
