--- website/src/routes/+layout.svelte
@@ -1,7 +1,0 @@
-<script>
-    import '../app.css'
-</script>
-
-<main>
-    <slot />
-</main>
