--- website/src/routes/+page.svelte
@@ -1,24 +1,0 @@
-<h1>Minecraft Crowd Control</h1>
-
-<p>
-    Minecraft <a href="https://crowdcontrol.live/">Crowd Control</a> is a Minecraft mod that allows your livestream
-    viewers on any platform to interact with your game using tips, bits, channel points, charitable donations, and more.
-    It can run as a client mod or as a server plugin.
-</p>
-
-<h2>Installation</h2>
-<p>
-    For help installing the mod, see the <a href="/setup">dynamic setup guide</a>.
-</p>
-
-<h2>Credits</h2>
-<p>
-    This mod was created by <a href="https://qixils.dev/">Lexi</a>
-    with the support of <a href="https://crowdcontrol.live">Crowd Control</a>.
-</p>
-
-<style>
-    #header {
-        text-align: center;
-    }
-</style>
