--- website/src/routes/setup/paper/automatic/+page.svelte
@@ -1,25 +1,0 @@
-<script lang="ts">
-    let version = "1.20.1";
-</script>
-
-<h1>Paper {version} Automatic Setup</h1>
-
-<p>The following steps provide a quick setup for running a local game server.</p>
-
-<ol>
-    <li>Download and install the <a href="https://crowdcontrol.live/">Crowd Control app</a>.</li>
-    <li>In the <strong>Game Library</strong> tab, select <strong>Minecraft</strong>.</li>
-    <li>Select <strong>Configure Minecraft</strong>.</li>
-    <li>Enter your Minecraft username and click next.</li>
-    <li>Select <strong>Paper</strong> and click next.</li>
-    <li>Select <strong>Local</strong>.</li>
-    <li>Select <strong>Click to install</strong> and <strong>Okay</strong>.</li>
-    <li>(Optional) You may now select <strong>Open Folder</strong> next to <strong>Minecraft</strong> on the <strong>Game Paths</strong> tab and copy any Paper plugins you want to play with into the <code>Paper/plugins</code> folder.</li>
-    <li>In the <strong>Live</strong> tab, select <strong>Launch Server</strong>.</li>
-    <li>Launch your instance of Minecraft {version}.</li>
-    <li>In Minecraft, add the server <code>localhost</code> to your multiplayer server list and join it.</li>
-    <li>(Optional) If you're sharing effects with another streamer, run the command <code>/account link INSERT_USERNAME</code> to receive their effects. The username should come from what's displayed in the top left corner of their Crowd Control app.</li>
-    <li>Select <strong>Start Session</strong> in the Crowd Control app.</li>
-</ol>
-
-<p>If you experience issues at any point, please reach out for help on our <a href="https://discord.gg/warpworld">Discord</a>.</p>
